Alter repo initialization pattern
Rather than performing all the exception handling in the constructor, I
noticed that it's not actually possible to always use the same
branch_name for all the checks (since the exercises may require
branching) and as a result, things like the user_commits may not always
be run against the same branch.

To address this, rather than focusing on performing all the
initializations and state management in the constructor, I have opted
for a more method-first approach - where all the necessary data is
accessible via method calls instead. This way, we can avoid encountering
complex errors BEFORE we actually need them