Adopt new exception handling pattern
Previously, when an autograding script encounters an exception (missing
start commit), the problem set PR comment would be an ambiguous
exception message that does not detail the issue. This can be painful to
debug and annoying to figure out.

This change does the following:

- Introduce custom exception types for both invalid states (like missing
the Git Mastery start commit) and wrong answers (the kind surfaced
during autograding)
- Expand @autograder() decorator to also handle exceptions raised to
return them as GitAutograderOutput objects that can be parsed by the
autograding workflow for the student to view
- Move the output saving behavior to the @autograder() decorator,
ensuring that we do not accidentally save outputs for local testing
(since they do not rely on the decorator's behavior)